GUI/requirements.txt
# Cornell Hyperloop GUI Dependencies
# Using latest compatible versions for Python 3.11+
# Pin only when specific versions are required for compatibility

# Core web framework
dash>=2.18.1
dash-bootstrap-components>=1.6.0
flask>=3.0.3

# Data processing (Python 3.11 compatible versions)
numpy>=2.0.0
pandas>=2.0.0
plotly>=5.24.0

# Serial communication
pyserial>=3.5

# Utility libraries
requests>=2.32.0
pyyaml>=6.0.0
python-dateutil>=2.9.0

# Core Python utilities (let pip handle these)
click
certifi
charset-normalizer
blinker
future
idna
importlib-metadata
iso8601
itsdangerous
jinja2
markupsafe
nest-asyncio
packaging
pytz
retrying
six
tenacity
typing-extensions
tzdata
urllib3
werkzeug
zipp

# Remove duplicate serial package
# serial==0.0.97  # This conflicts with pyserial
